refactor generator, simplicity test to BigInteger

+++ src/main/java/com/example/rsalab/util/math/PrimeNumberGenerator.java
@@ -2,12 +2,13 @@ package com.example.rsalab.util.math;
 
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
 import java.util.function.Supplier;
 
 @Component
-public class PrimeNumberGenerator implements Supplier<Long> {
+public class PrimeNumberGenerator implements Supplier<BigInteger> {
     @Override
-    public Long get() {
+    public BigInteger get() {
         return null;
     }
 }
